requirements.txt
flask>=2.2.0,<3
python-dotenv>=0.19.0
gunicorn>=21.2.0
